+++ lib/chart/render/decorations/selected_item_decoration.dart
@@ -9,10 +9,12 @@ class SelectedItemDecoration extends DecorationPainter {
     this.backgroundColor = Colors.grey,
     this.selectedStyle = const TextStyle(fontSize: 13.0),
     this.animate = false,
-    this.showText = true,
+    bool showText = true,
     this.showOnTop = true,
     this.selectedArrayIndex = 0,
-  });
+    this.child,
+    this.topMargin = 0.0,
+  }) : showText = showText && (child == null);
 
   /// Index of selected item
   final int? selectedItem;
@@ -33,21 +35,25 @@ class SelectedItemDecoration extends DecorationPainter {
   /// Color of selected item text
   final TextStyle selectedStyle;
 
+  final Widget? child;
+  final double topMargin;
+
   /// Index of list whose data to show
   /// By default it will use first list to this value will be `0`
   final int selectedArrayIndex;
 
-  // @override
-  // Widget getRenderer(ChartState state) {
-  //   return ChartDecorationChildRenderer(
-  //       state,
-  //       this,
-  //       Container(
-  //         width: 100.0,
-  //         height: 100.0,
-  //         color: Colors.red,
-  //       ));
-  // }
+  @override
+  Widget getRenderer(ChartState state) {
+    if (child != null) {
+      return ChartDecorationChildRenderer(
+        state,
+        this,
+        child!,
+      );
+    }
+
+    return super.getRenderer(state);
+  }
 
   @override
   void initDecoration(ChartState state) {
@@ -59,8 +65,7 @@ class SelectedItemDecoration extends DecorationPainter {
   @override
   Size layoutSize(BoxConstraints constraints, ChartState state) {
     final _listSize = state.data.listSize;
-    final _itemWidth =
-        constraints.deflate(state.defaultMargin).maxWidth / _listSize;
+    final _itemWidth = constraints.deflate(state.defaultMargin).maxWidth / _listSize;
 
     return Size(
       _itemWidth,
@@ -70,20 +75,28 @@ class SelectedItemDecoration extends DecorationPainter {
 
   @override
   Offset applyPaintTransform(ChartState state, Size size) {
-    final _width =
-        (size.width - state.defaultMargin.horizontal) / state.data.listSize;
+    final _width = (size.width - state.defaultMargin.horizontal) / state.data.listSize;
 
     final selectedItem = this.selectedItem;
     if (selectedItem != null) {
-      return Offset(state.defaultMargin.left + _width * selectedItem,
-          (state.defaultMargin - marginNeeded()).top);
+      final _selectedItemMax = state.data.items[selectedArrayIndex][selectedItem].max ?? 0.0;
+      final _selectedItemMin = state.data.items[selectedArrayIndex][selectedItem].min ?? 0.0;
+
+      final _maxValue = state.data.maxValue - state.data.minValue;
+      final _height = size.height - marginNeeded().vertical;
+      final scale = _height / _maxValue;
+
+      return Offset(
+          state.defaultMargin.left + _width * selectedItem,
+          showOnTop
+              ? (state.defaultMargin - marginNeeded()).top
+              : size.height - ((_selectedItemMax - (min(_selectedItemMin, state.data.minValue))) * scale));
     }
 
     return Offset.zero;
   }
 
-  void _drawText(
-      Canvas canvas, Size size, double totalWidth, ChartState state) {
+  void _drawText(Canvas canvas, Size size, double totalWidth, ChartState state) {
     final _item = selectedItem;
     if (_item == null) {
       return;
@@ -97,10 +110,8 @@ class SelectedItemDecoration extends DecorationPainter {
       selectedStyle,
       hasMaxWidth: false,
     );
-    final _itemWidth = max(
-        state.itemOptions.minBarWidth ?? 0.0,
-        min(state.itemOptions.maxBarWidth ?? double.infinity,
-            size.width - state.itemOptions.padding.horizontal));
+    final _itemWidth = max(state.itemOptions.minBarWidth ?? 0.0,
+        min(state.itemOptions.maxBarWidth ?? double.infinity, size.width - state.itemOptions.padding.horizontal));
 
     const _size = 2.0;
     final _maxValue = state.data.maxValue - state.data.minValue;
@@ -132,21 +143,7 @@ class SelectedItemDecoration extends DecorationPainter {
     }
 
     if (showOnTop) {
-      canvas.drawRect(
-        Rect.fromPoints(
-          Offset(
-            state.itemOptions.padding.left + _itemWidth / 2 - _size / 2,
-            (selectedStyle.fontSize ?? 0.0) * 1.4,
-          ),
-          Offset(
-            state.itemOptions.padding.left + _itemWidth / 2 + _size / 2,
-            size.height -
-                ((_itemMaxValue - (min(_itemMinValue, state.data.minValue))) *
-                    scale),
-          ),
-        ),
-        Paint()..color = selectedColor,
-      );
+      _drawLine(canvas, size, state);
     }
 
     canvas.drawRRect(
@@ -155,26 +152,18 @@ class SelectedItemDecoration extends DecorationPainter {
               Offset(
                 width / 2 - _maxValuePainter.width / 2,
                 ((size.height -
-                            ((_itemMaxValue -
-                                    (min(_itemMinValue, state.data.minValue))) *
-                                scale) -
+                            ((_itemMaxValue - (min(_itemMinValue, state.data.minValue))) * scale) -
                             (selectedStyle.fontSize ?? 0.0) * 1.4) *
                         (showOnTop ? 0 : 1)) +
-                    ((showOnTop ? 1 : 0) *
-                        (selectedStyle.fontSize ?? 0.0) *
-                        1.4),
+                    ((showOnTop ? 1 : 0) * (selectedStyle.fontSize ?? 0.0) * 1.4),
               ),
               Offset(
                 width / 2 + _maxValuePainter.width / 2,
                 ((size.height -
-                            ((_itemMaxValue -
-                                    (min(_itemMinValue, state.data.minValue))) *
-                                scale) -
+                            ((_itemMaxValue - (min(_itemMinValue, state.data.minValue))) * scale) -
                             (selectedStyle.fontSize ?? 0.0) * 0.4) *
                         (showOnTop ? 0 : 1)) +
-                    ((showOnTop ? 1 : 0) *
-                        (selectedStyle.fontSize ?? 0.0) *
-                        0.4),
+                    ((showOnTop ? 1 : 0) * (selectedStyle.fontSize ?? 0.0) * 0.4),
               )),
           const Radius.circular(8.0),
         ).inflate(4),
@@ -185,9 +174,7 @@ class SelectedItemDecoration extends DecorationPainter {
       Offset(
         width / 2 - _maxValuePainter.width / 2,
         ((size.height -
-                    ((_itemMaxValue -
-                            (min(_itemMinValue, state.data.minValue))) *
-                        scale) -
+                    ((_itemMaxValue - (min(_itemMinValue, state.data.minValue))) * scale) -
                     (selectedStyle.fontSize ?? 0.0) * 1.4) *
                 (showOnTop ? 0 : 1)) +
             ((showOnTop ? 1 : 0) * (selectedStyle.fontSize ?? 0.0) * 0.4),
@@ -195,8 +182,45 @@ class SelectedItemDecoration extends DecorationPainter {
     );
   }
 
+  void _drawLine(Canvas canvas, Size size, ChartState state) {
+    final _item = selectedItem;
+    if (_item == null) {
+      return;
+    }
+    final _selectedItem = state.data.items[selectedArrayIndex][_item];
+
+    final _itemWidth = max(state.itemOptions.minBarWidth ?? 0.0,
+        min(state.itemOptions.maxBarWidth ?? double.infinity, size.width - state.itemOptions.padding.horizontal));
+
+    const _size = 2.0;
+    final _maxValue = state.data.maxValue - state.data.minValue;
+    final _height = size.height - marginNeeded().vertical;
+    final scale = _height / _maxValue;
+
+    final _itemMaxValue = _selectedItem.max ?? 0.0;
+    final _itemMinValue = _selectedItem.min ?? 0.0;
+
+    canvas.drawRect(
+      Rect.fromPoints(
+        Offset(
+          state.itemOptions.padding.left + _itemWidth / 2 - _size / 2,
+          (selectedStyle.fontSize ?? 0.0) * 1.4,
+        ),
+        Offset(
+          state.itemOptions.padding.left + _itemWidth / 2 + _size / 2,
+          size.height - ((_itemMaxValue - (min(_itemMinValue, state.data.minValue))) * scale),
+        ),
+      ),
+      Paint()..color = selectedColor,
+    );
+  }
+
   @override
   void draw(Canvas canvas, Size size, ChartState state) {
+    if (child != null) {
+      return;
+    }
+
     final _listSize = state.data.listSize;
     final _item = selectedItem;
 
@@ -214,8 +238,7 @@ class SelectedItemDecoration extends DecorationPainter {
 
   void _drawItem(Canvas canvas, Size size, ChartState state) {
     canvas.drawRect(
-      Rect.fromPoints(
-          Offset(0.0, marginNeeded().top), Offset(size.width, size.height)),
+      Rect.fromPoints(Offset(0.0, marginNeeded().top), Offset(size.width, size.height)),
       Paint()
         ..color = backgroundColor
         ..blendMode = BlendMode.overlay,
@@ -225,7 +248,13 @@ class SelectedItemDecoration extends DecorationPainter {
   @override
   EdgeInsets marginNeeded() {
     return EdgeInsets.only(
-        top: showText && showOnTop ? (selectedStyle.fontSize ?? 0) * 1.8 : 0.0);
+        top: showOnTop
+            ? (showText
+                ? (selectedStyle.fontSize ?? 0) * 1.8
+                : child != null
+                    ? topMargin
+                    : 0.0)
+            : 0.0);
   }
 
   @override
@@ -233,21 +262,17 @@ class SelectedItemDecoration extends DecorationPainter {
     if (endValue is SelectedItemDecoration) {
       return SelectedItemDecoration(
         animate
-            ? (lerpDouble(selectedItem?.toDouble(),
-                        endValue.selectedItem?.toDouble(), t) ??
-                    0)
-                .round()
+            ? (lerpDouble(selectedItem?.toDouble(), endValue.selectedItem?.toDouble(), t) ?? 0).round()
             : endValue.selectedItem,
-        selectedColor: Color.lerp(selectedColor, endValue.selectedColor, t) ??
-            endValue.selectedColor,
-        backgroundColor:
-            Color.lerp(backgroundColor, endValue.backgroundColor, t) ??
-                endValue.backgroundColor,
-        selectedStyle:
-            TextStyle.lerp(selectedStyle, endValue.selectedStyle, t) ??
-                endValue.selectedStyle,
+        selectedColor: Color.lerp(selectedColor, endValue.selectedColor, t) ?? endValue.selectedColor,
+        backgroundColor: Color.lerp(backgroundColor, endValue.backgroundColor, t) ?? endValue.backgroundColor,
+        selectedStyle: TextStyle.lerp(selectedStyle, endValue.selectedStyle, t) ?? endValue.selectedStyle,
         animate: endValue.animate,
         selectedArrayIndex: endValue.selectedArrayIndex,
+        showOnTop: t < 0.5 ? showOnTop : endValue.showOnTop,
+        showText: t < 0.5 ? showText : endValue.showText,
+        child: t < 0.5 ? child : endValue.child,
+        topMargin: lerpDouble(topMargin, endValue.topMargin, t) ?? 0.0,
       );
     }